Retrieve authenticator schemas for all authenticators
